+++ src/main/java/com/Bridge/bridge/dto/response/apple/AppleResponse.java
@@ -1,4 +1,4 @@
-package com.Bridge.bridge.dto.response;
+package com.Bridge.bridge.dto.response.apple;
 
 import lombok.Data;
