+++ src/main/java/com/Bridge/bridge/dto/response/apple/AppleTokenResponse.java
@@ -1,4 +1,4 @@
-package com.Bridge.bridge.security;
+package com.Bridge.bridge.dto.response.apple;
 
 import lombok.Data;
